requirements.txt
apache-airflow-providers-docker
docker
jupyter
